requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
requests==2.31.0
pysam==0.19.1
PyVCF==0.6.8
pandas==1.5.3
openpyxl==3.1.2
hashable==0.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1